refactor: decompose run-oxlint.ts + inspect.ts, adopt Layer.mock in tests
Three behavior-preserving structural wins, all validated by the
existing 1448-test suite.

**run-oxlint.ts: 673 → 244 lines (-64%)**

Extracted five focused modules under `runners/oxlint/`:

- `spawn-oxlint.ts` — one subprocess invocation, hard ceilings on
  wall time and output bytes, tagged-error reject contract.
- `parse-output.ts` — oxlint stdout → `Diagnostic[]`, including the
  leading-notice-line tolerance and category/recommendation lookups.
- `resolve-paths.ts` — oxlint binary, plugin path, tsconfig path.
- `validate-rule-registration.ts` — one-time metadata-drift check.
- `spawn-batches.ts` — the batching loop with binary-split retry and
  per-batch dropped-file tracking.

`run-oxlint.ts` is now a slim orchestrator: it resolves
plugins/extends/ignore patterns, builds the oxlintrc.json, neutralizes
inline disable directives in audit mode, runs the batches, and on
extends-related crashes retries once with extends stripped (still
carrying `userPlugins` through to the fallback config).

**inspect.ts: 551 → 520 lines, layer wiring extracted**

`cli/utils/build-runtime-layers.ts` (82 lines) owns the production
`Layer.mergeAll(...)` stack. `inspect.ts` keeps its post-scan
rendering responsibilities; layer composition is no longer mixed in
with `Console.log` calls and spinner finalization.

**Tests: Layer.mock for one-off service overrides**

Replaced `Layer.succeed(Service, Service.of({...all methods...}))`
with `Layer.mock(Service, { ...only methods this test exercises... })`
in:

- `tests/run-inspect.test.ts` (Project, Linter, DeadCode overrides)
- `tests/services/staged-files.test.ts` (Git override × 2)

Dropped the `Git.of({...})` ceremony and the `Context.Tag.Service`
casts — `Layer.mock` raises an unimplemented defect if a test
exercises a method the mock didn't provide, which is the failure
shape we want anyway.

No behavior change. Full validation pass: typecheck + lint + format +
smoke + 1448 tests all green.
